+++ packages/zettel/zettel-editor/.prettierrc.json
@@ -1,0 +1,3 @@
+{
+  "printWidth": 80
+}
