+++ packages/zettel/zettel-editor/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "@opral/zettel-editor",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "exports": {
+    ".": "./dist/index.js"
+  },
+  "scripts": {
+    "dev": "tsc --watch",
+    "build": "tsc",
+    "test": "vitest run",
+    "test:watch": "vitest",
+    "build:storybook": "storybook build",
+    "format": "prettier ./src --write",
+    "storybook": "storybook dev -p 6006"
+  },
+  "dependencies": {
+    "@opral/zettel-ast": "workspace:*",
+    "@opral/zettel-html": "workspace:*",
+    "@opral/zettel-lexical": "workspace:*",
+    "lexical": "^0.30.0",
+    "lit": "^3.2.1"
+  },
+  "devDependencies": {
+    "@open-wc/lit-helpers": "^0.7.0",
+    "@storybook/addon-actions": "^8.6.12",
+    "@storybook/addon-essentials": "8.6.12",
+    "@storybook/blocks": "8.6.12",
+    "@storybook/preview-api": "8.6.12",
+    "@storybook/test": "8.6.12",
+    "@storybook/web-components": "8.6.12",
+    "@storybook/web-components-vite": "8.6.12",
+    "prettier": "^3.5.3",
+    "storybook": "8.6.12",
+    "typescript": "~5.7.2",
+    "vite": "^6.3.0",
+    "vitest": "^3.1.1"
+  }
+}
