+++ packages/zettel/zettel-editor/tsconfig.json
@@ -1,0 +1,29 @@
+{
+  "include": ["src"],
+  "exclude": ["**/*.stories.ts"],
+  "compilerOptions": {
+    "outDir": "./dist",
+    "sourceMap": true,
+    "declarationMap": true,
+    "target": "ES2022",
+    "experimentalDecorators": false,
+    "useDefineForClassFields": false,
+    "module": "Node16",
+    "lib": ["ES2022", "DOM", "DOM.Iterable"],
+    "skipLibCheck": true,
+
+    /* Bundler mode */
+    "moduleResolution": "node16",
+    "isolatedModules": true,
+    "moduleDetection": "force",
+
+    "declaration": true,
+
+    /* Linting */
+    "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noFallthroughCasesInSwitch": true,
+    "noUncheckedSideEffectImports": true
+  }
+}
