+++ packages/zettel/zettel-lexical/package.json
@@ -11,36 +11,22 @@
     "build": "tsc",
     "test": "vitest run",
     "test:watch": "vitest",
-    "build:storybook": "storybook build",
-    "format": "prettier ./src --write",
-    "storybook": "storybook dev -p 6006"
+    "format": "prettier ./src --write"
   },
   "dependencies": {
-    "@lexical/clipboard": "^0.30.0",
-    "@lexical/history": "^0.30.0",
-    "@lexical/plain-text": "^0.30.0",
-    "@lexical/rich-text": "^0.30.0",
-    "@lexical/selection": "^0.30.0",
     "@lexical/utils": "^0.30.0",
     "@opral/zettel-ast": "workspace:*",
-    "@opral/zettel-html": "workspace:*",
-    "lexical": "^0.30.0",
-    "lit": "^3.2.1"
+    "@opral/zettel-html": "workspace:*"
+  },
+  "peerDependencies": {
+    "lexical": "^0.30.0"
   },
   "devDependencies": {
-    "@lexical/headless": "^0.30.0",
     "@open-wc/lit-helpers": "^0.7.0",
-    "@storybook/addon-actions": "^8.6.12",
-    "@storybook/addon-essentials": "8.6.12",
-    "@storybook/blocks": "8.6.12",
-    "@storybook/preview-api": "8.6.12",
-    "@storybook/test": "8.6.12",
-    "@storybook/web-components": "8.6.12",
-    "@storybook/web-components-vite": "8.6.12",
     "@types/jsdom": "^21.1.7",
     "jsdom": "^26.1.0",
     "prettier": "^3.5.3",
-    "storybook": "8.6.12",
+    "lexical": "^0.30.0",
     "typescript": "~5.7.2",
     "vite": "^6.3.0",
     "vitest": "^3.1.1"
