+++ packages/zettel/zettel-lexical/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./nodes.js";
+export { registerZettelLexicalPlugin } from "./plugin.js";
+export { toLexicalState, fromLexicalState } from "./parse-serialize.js";
